REMOVE: Downloader class - inline it back.
